Move Provider into own directory
