+++ __tests__/helpers/BadgeContext.test.tsx
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, it, jest } from "@jest/globals";
 
-import { updateBadgeState } from "#/helpers/BadgeContext";
 import BadgeStore from "#/helpers/Stores/BadgeStore";
+import { updateBadgeState } from "#/helpers/provider/BadgeProvider";
 
 // Mock the BadgeStore
 jest.mock("../../src/helpers/Stores/BadgeStore", () => ({
