+++ src/helpers/Stores/BadgeStore.ts
@@ -1,5 +1,5 @@
-import { BadgeState } from "#/helpers/BadgeContext";
 import BaseStore from "#/helpers/Storage";
+import { BadgeState } from "#/helpers/provider/BadgeProvider";
 
 /**
  * Class for managing badge-related state.
